finish nodes highlighting
